fix db, fix ban bot from user

+++ app/models/message.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Boolean, Integer, ForeignKey, Text
+from sqlalchemy import Boolean, ForeignKey, Text, BigInteger
 from sqlalchemy.orm import mapped_column, relationship
 
 from app.core.db import Base
@@ -6,9 +6,9 @@ from app.core.db import Base
 
 class Message(Base):
     """Модель сообщений"""
-    telegram_user_id = mapped_column(Integer, ForeignKey(
+    telegram_user_id = mapped_column(BigInteger, ForeignKey(
         'user.telegram_id'), nullable=False)
-    answer_to_user_id = mapped_column(Integer, ForeignKey(
+    answer_to_user_id = mapped_column(BigInteger, ForeignKey(
         'user.telegram_id'), nullable=True)
     text = mapped_column(Text, nullable=True)
     attachments = mapped_column(Boolean, default=False)
@@ -16,3 +16,4 @@ class Message(Base):
         telegram_user_id], lazy='subquery')
     answer_to_user = relationship('User', backref='answers', foreign_keys=[
         answer_to_user_id])
+
